requirements.txt
Flask==3.0.0
PyMuPDF==1.23.8
python-dotenv==1.0.0
gunicorn==21.2.0 